enhance the episode and season
First thing to do in a season is to fill the agent's memory. The agent
has wrapper methods to call various memory methods to not break the Law
of Demeter. Now running an episode or season will take two parameters.
One for rendering the game and another for telling the game master if
it's in training mode or not. Also break up fill_agent_memory and
render_game into their own methods. render_game is crudely slowing down
the rendering to actually see the game in real time. Frames are being
track of by the game master and are different from the moves taken by
the agent.